Improve pyproject.toml somewhat
this was an attempt to clean up what's contained in the wheel and in the
sdist. However, setuptools is just broken so it seems hopeless to
actually make this work out.

I tried to use meson but we cannot work around the namespace package
foo in SageMath that needs Cython 3.